Handle surface creation attempt while stopped
In the case where the application is in the Android stopped state,
surfaces will never be created. Due to the timing of the multiple threads
we sometimes get into this state. In this case, we will now bail out of the
wait condition rather than blocking until the application changes state.